# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Client Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      responses:
        "200":
          description: All courses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CourseSummary" }
    post:
      operationId: createCourse
      tags: [courses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCourseBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
  /courses/{courseId}:
    parameters:
      - { name: courseId, in: path, required: true, schema: { type: string } }
    get:
      operationId: getCourse
      tags: [courses]
      summary: Full course tree (chapters + lessons) with user progress
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CourseDetail" }
    patch:
      operationId: updateCourse
      tags: [courses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCourseBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Course" }
    delete:
      operationId: deleteCourse
      tags: [courses]
      responses:
        "204": { description: Deleted }
  /courses/{courseId}/chapters:
    parameters:
      - { name: courseId, in: path, required: true, schema: { type: string } }
    post:
      operationId: createChapter
      tags: [chapters]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateChapterBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chapter" }
  /chapters/{chapterId}:
    parameters:
      - { name: chapterId, in: path, required: true, schema: { type: string } }
    patch:
      operationId: updateChapter
      tags: [chapters]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateChapterBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Chapter" }
    delete:
      operationId: deleteChapter
      tags: [chapters]
      responses:
        "204": { description: Deleted }
...
  /dfy/projects:
    get:
      operationId: listMyProjects
      tags: [dfy]
      summary: Done-for-you projects visible to current user
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DfyProject" }
    post:
      operationId: createProject
      tags: [dfy]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProjectBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DfyProject" }
  /dfy/projects/{projectId}:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string } }
    get:
      operationId: getProject
      tags: [dfy]
      responses:
        "200":
          description: Project detail with updates
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DfyProjectDetail" }
    patch:
      operationId: updateProject
      tags: [dfy]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProjectBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DfyProject" }
  /dfy/projects/{projectId}/updates:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string } }
    post:
      operationId: addProjectUpdate
      tags: [dfy]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProjectUpdateBody" }
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectUpdate" }
  /dfy/projects/{projectId}/milestones:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string } }
    post:
      operationId: addProjectMilestone
      tags: [dfy]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMilestoneBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectMilestone" }
  /dfy/milestones/{milestoneId}:
    parameters:
      - { name: milestoneId, in: path, required: true, schema: { type: string } }
    patch:
      operationId: updateProjectMilestone
      tags: [dfy]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMilestoneBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectMilestone" }
    delete:
      operationId: deleteProjectMilestone
      tags: [dfy]
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    CourseSummary:
      type: object
      required: [id, title, totalLessons]
      properties:
        id: { type: string }
        title: { type: string }
        subtitle: { type: string, nullable: true }
        coverImageUrl: { type: string, nullable: true }
        totalLessons: { type: integer }
        published: { type: boolean }
    Course:
      type: object
      required: [id, title, published, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        subtitle: { type: string, nullable: true }
        description: { type: string, nullable: true }
        coverImageUrl: { type: string, nullable: true }
        published: { type: boolean }
        createdAt: { type: string, format: date-time }
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          required: [chapters, completedLessonIds]
          properties:
            chapters:
              type: array
              items: { $ref: "#/components/schemas/ChapterWithLessons" }
            completedLessonIds:
              type: array
              items: { type: string }
    CreateCourseBody:
      type: object
      required: [title]
      properties:
        title: { type: string }
        subtitle: { type: string, nullable: true }
        description: { type: string, nullable: true }
        coverImageUrl: { type: string, nullable: true }
    UpdateCourseBody:
      type: object
      properties:
        title: { type: string }
        subtitle: { type: string, nullable: true }
        description: { type: string, nullable: true }
        coverImageUrl: { type: string, nullable: true }
        published: { type: boolean }
    Chapter:
      type: object
      required: [id, courseId, title, position]
      properties:
        id: { type: string }
        courseId: { type: string }
        title: { type: string }
        position: { type: integer }
    ChapterWithLessons:
      allOf:
        - $ref: "#/components/schemas/Chapter"
        - type: object
          required: [lessons]
          properties:
            lessons:
              type: array
              items: { $ref: "#/components/schemas/Lesson" }
    CreateChapterBody:
      type: object
      required: [title]
      properties:
        title: { type: string }
    UpdateChapterBody:
      type: object
      properties:
        title: { type: string }
        position: { type: integer }
    Lesson:
      type: object
      required: [id, chapterId, title, position, lessonType]
      properties:
        id: { type: string }
        chapterId: { type: string }
        title: { type: string }
        position: { type: integer }
        lessonType:
          type: string
          enum: [video, text]
        videoUrl: { type: string, nullable: true }
        durationMinutes: { type: integer, nullable: true }
        published: { type: boolean }
...
    DfyProject:
      type: object
      required: [id, clientUserId, title, status, createdAt]
      properties:
        id: { type: string }
        clientUserId: { type: string }
        clientName: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        status:
          type: string
          enum: [discovery, in_progress, review, delivered, on_hold]
        progressPct: { type: integer }
        nextMilestone: { type: string, nullable: true }
        nextMilestoneDate: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    DfyProjectDetail:
      allOf:
        - $ref: "#/components/schemas/DfyProject"
        - type: object
          required: [updates, milestones]
          properties:
            updates:
              type: array
              items: { $ref: "#/components/schemas/ProjectUpdate" }
            milestones:
              type: array
              items: { $ref: "#/components/schemas/ProjectMilestone" }
    ProjectUpdate:
      type: object
      required: [id, projectId, authorName, body, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        authorName: { type: string }
        body: { type: string }
        createdAt: { type: string, format: date-time }
    ProjectMilestone:
      type: object
      required: [id, title, status]
      properties:
        id: { type: string }
        title: { type: string }
        status:
          type: string
          enum: [pending, done]
        dueDate: { type: string, format: date-time, nullable: true }
    CreateMilestoneBody:
      type: object
      required: [title]
      properties:
        title: { type: string }
        dueDate: { type: string, format: date-time, nullable: true }
    UpdateMilestoneBody:
      type: object
      properties:
        title: { type: string }
        status: { type: string, enum: [pending, done] }
        dueDate: { type: string, format: date-time, nullable: true }
    CreateProjectBody:
      type: object
      required: [title, clientUserId]
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        clientUserId: { type: string }
        status:
          type: string
          enum: [discovery, in_progress, review, delivered, on_hold]
    UpdateProjectBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        status:
          type: string
          enum: [discovery, in_progress, review, delivered, on_hold]
        progressPct: { type: integer }
        nextMilestone: { type: string, nullable: true }
        nextMilestoneDate: { type: string, format: date-time, nullable: true }
    CreateProjectUpdateBody:
      type: object
      required: [body]
      properties:
        body: { type: string }
...